RavenDB-18529 Sharding - Tasks List View - RavenDB ETL - reuse for olap & sql
