+++ src/Raven.Studio/typescript/components/pages/database/tasks/list/panels/RavenEtlPanel.tsx
@@ -3,44 +3,24 @@ import { useAccessManager } from "hooks/useAccessManager";
 import { RichPanel, RichPanelDetailItem, RichPanelDetails, RichPanelHeader } from "../../../../../common/RichPanel";
 import {
     ConnectionStringItem,
+    EmptyScriptsWarning,
     ICanShowTransformationScriptPreview,
     OngoingTaskActions,
     OngoingTaskName,
     OngoingTaskStatus,
 } from "../shared";
 import { useAppUrls } from "hooks/useAppUrls";
-import { OngoingTaskInfo, OngoingTaskRavenEtlInfo } from "../../../../../models/tasks";
+import { OngoingTaskRavenEtlInfo } from "../../../../../models/tasks";
 import { BaseOngoingTaskPanelProps, useTasksOperations } from "../shared";
 import { OngoingTaskDistribution } from "./OngoingTaskDistribution";
 
 type RavenEtlPanelProps = BaseOngoingTaskPanelProps<OngoingTaskRavenEtlInfo>;
-
-function findScriptsWithOutMatchingDocuments(data: OngoingTaskInfo): string[] {
-    const perScriptCounts = new Map<string, number>();
-    data.nodesInfo.forEach((node) => {
-        if (node.progress) {
-            node.progress.forEach((progress) => {
-                const transformationName = progress.transformationName;
-                perScriptCounts.set(
-                    transformationName,
-                    (perScriptCounts.get(transformationName) ?? 0) + progress.global.total
-                );
-            });
-        }
-    });
-
-    return Array.from(perScriptCounts.entries())
-        .filter((x) => x[1] === 0)
-        .map((x) => x[0]);
-}
 
 function Details(props: RavenEtlPanelProps & { canEdit: boolean }) {
     const { data, canEdit, db } = props;
     const connectionStringDefined = !!data.shared.destinationDatabase;
     const { appUrl } = useAppUrls();
     const connectionStringsUrl = appUrl.forConnectionStrings(db, "ravendb", data.shared.connectionStringName);
-
-    const emptyScripts = findScriptsWithOutMatchingDocuments(data);
 
     return (
         <RichPanelDetails>
@@ -74,14 +54,7 @@ function Details(props: RavenEtlPanelProps & { canEdit: boolean }) {
                     <div className="value">{data.shared.topologyDiscoveryUrls.join(", ")}</div>
                 </RichPanelDetailItem>
             )}
-            {emptyScripts.length > 0 && (
-                <RichPanelDetailItem className="text-warning">
-                    <small>
-                        <i className="icon-warning" />
-                        Following scripts don't match any documents: {emptyScripts.join(", ")}
-                    </small>
-                </RichPanelDetailItem>
-            )}
+            <EmptyScriptsWarning task={data} />
         </RichPanelDetails>
     );
 }
